+++ tests/Core/ProjectTest.php
@@ -1,0 +1,209 @@
+<?php
+
+namespace Leedch\Codemonkey\Core;
+
+use Leedch\Codemonkey\Core\Project;
+use PHPUnit\Framework\TestCase;
+
+class ProjectTest extends TestCase {
+
+    protected function getConfigArray()
+    {
+        $expected = [
+            "projectname" => "TestCase",
+            "files" => [
+                [
+                    "name" => "tests/testcase.php",
+                    "templates" => [
+                        "../vendor/leedch/codemonkey/templates/testcase.php.txt"
+                    ],
+                    "attributes" => [
+                        "testClassName" => "testcase",
+                        "namespace" => "Testclasses\\Testcases",
+                        "testClassAttributes" => 'protected $home = "home";'."\n".'protected $isActive = true;'."\n",
+                        "testClassMethods" => "public function __construct() \n{\n}\n"
+                    ]
+                ],
+                [
+                    "name" => "tests/testcase2.php",
+                    "templates" => [
+                        "../vendor/leedch/codemonkey/templates/testcase.php.txt"
+                    ],
+                    "attributes" => [
+                        "testClassName" => "testcase2",
+                        "namespace" => "Testclasses\\Testcases",
+                        "testClassAttributes" => 'protected $home = "home2";'."\n".'protected $isActive = false;'."\n",
+                        "testClassMethods" => "public function __construct() \n{\n}\n"
+                    ]
+                ]
+            ],
+            "attributes" => [
+                [
+                    "label" => "Name of first Class",
+                    "name" => "classname1",
+                    "default" => "FirstClass"
+                ],
+                [
+                    "label" => "Name of first Class attribute",
+                    "name" => "input1",
+                    "default" => "devilsAttribute"
+                ],
+                [
+                    "label" => "Namespace for Class 1",
+                    "name" => "namespace1",
+                    "default" => "Leedch\\Codemonkey\\Demo2"
+                ]
+            ]
+        ];
+        return $expected;
+    }
+    
+    public function testLoadConfig() 
+    {
+        
+        $m = new Project();
+        $pathConfigFile = __DIR__."/../../examples/testcase.json";
+        $m->loadConfig($pathConfigFile);
+        $expected = $this->getConfigArray();
+        $this->assertAttributeEquals($expected, 'config', $m);
+    }
+
+    public function testLoadConfigJson() 
+    {
+        $m = new Project();
+        $Json = file_get_contents(__DIR__."/../../examples/testcase.json");
+        $m->loadConfigJson($Json);
+        $expected = $this->getConfigArray();
+        $this->assertAttributeEquals($expected, 'config', $m);
+    }
+
+    public function testExecute() 
+    {
+        $m = new Project();
+        $pathConfigFile = __DIR__."/../../examples/testcase.json";
+        $m->loadConfig($pathConfigFile);
+        $_SERVER['REQUEST_URI'] = '/';
+        ob_start();
+        $m->execute();
+        $response = ob_get_clean();
+        
+        $expected = "<?xml version=\"1.0\" encoding=\"utf-8\" ?>\n"
+                    ."<!DOCTYPE html>\n"
+                    ."<html lang=\"en\">\n"
+                    ."<head>\n"
+                    ."<title>\n"
+                    ."Codemonkey</title>\n"
+                    ."</head>\n"
+                    ."<body>\n"
+                    ."<form action=\"/\" method=\"post\">\n"
+                    ."<table>\n"
+                    ."<tbody>\n"
+                    ."<tr>\n"
+                    ."<td>\n"
+                    ."<label for=\"classname1\">\n"
+                    ."Name of first Class</label>\n"
+                    ."</td>\n"
+                    ."<td>\n"
+                    ."<input name=\"classname1\" type=\"input\" value=\"FirstClass\" />\n"
+                    ."</td>\n"
+                    ."</tr>\n"
+                    ."<tr>\n"
+                    ."<td>\n"
+                    ."<label for=\"input1\">\n"
+                    ."Name of first Class attribute</label>\n"
+                    ."</td>\n"
+                    ."<td>\n"
+                    ."<input name=\"input1\" type=\"input\" value=\"devilsAttribute\" />\n"
+                    ."</td>\n"
+                    ."</tr>\n"
+                    ."<tr>\n"
+                    ."<td>\n"
+                    ."<label for=\"namespace1\">\n"
+                    ."Namespace for Class 1</label>\n"
+                    ."</td>\n"
+                    ."<td>\n"
+                    ."<input name=\"namespace1\" type=\"input\" value=\"Leedch\Codemonkey\Demo2\" />\n"
+                    ."</td>\n"
+                    ."</tr>\n"
+                    ."</tbody>\n"
+                    ."</table>\n"
+                    ."<button type=\"submit\">\n"
+                    ."Generate Code</button>\n"
+                    ."</form>\n"
+                    ."</body>\n"
+                    ."</html>\n";
+            
+        $this->assertEquals($expected, $response);
+    }
+
+    public function testCreateFiles() 
+    {
+        $m = new Project();
+        $pathConfigFile = __DIR__."/../../examples/testcase.json";
+        $m->loadConfig($pathConfigFile);
+        $_POST['classname1'] = 'testinput';
+        $_POST['input1'] = 'testinput';
+        $_POST['namespace'] = 'namespace Test\\Namespace';
+        
+        $m->createFiles();
+        $file1 = 'tests/testcase.php';
+        $file2 = 'tests/testcase2.php';
+        $this->assertFileExists(codemonkey_pathTempDir . $file1);
+        $this->assertGreaterThan(0, filesize(codemonkey_pathTempDir . $file1));
+        $this->assertFileExists(codemonkey_pathTempDir . $file2);
+        $file1Content = file_get_contents(codemonkey_pathTempDir . $file1);
+        $file1ExpectedContent = "<?php\n"
+            ."\n"
+            ."Testclasses\Testcases\n"
+            ."\n"
+            ."class testcase {\n"
+            .'protected $home = "home";'."\n"
+            .'protected $isActive = true;'."\n"
+            ."\n"
+            ."public function __construct() \n"
+            ."{\n"
+            ."}\n"
+            ."\n"
+            ."}\n\n";
+        $this->assertEquals($file1ExpectedContent, $file1Content);
+        $this->assertGreaterThan(0, filesize(codemonkey_pathTempDir . $file2));
+        $file2Content = file_get_contents(codemonkey_pathTempDir . $file2);
+        $file2ExpectedContent = "<?php\n"
+            ."\n"
+            ."Testclasses\Testcases\n"
+            ."\n"
+            ."class testcase2 {\n"
+            .'protected $home = "home2";'."\n"
+            .'protected $isActive = false;'."\n"
+            ."\n"
+            ."public function __construct() \n"
+            ."{\n"
+            ."}\n"
+            ."\n"
+            ."}\n\n";
+        $this->assertEquals($file2ExpectedContent, $file2Content);
+    }
+
+    public function testClearTempFolder() 
+    {
+        $m = new Project();
+        $m->clearTempFolder();
+        $this->assertDirectoryExists(codemonkey_pathTempDir);
+        $arrContents = scandir(codemonkey_pathTempDir);
+        $expectedContent = ['.','..'];
+        $this->assertEquals($expectedContent, $arrContents);
+    }
+
+    public function testReturnZipFile() 
+    {
+        //Skip this test, as it returns headers in HTML and cant do that while
+        //PHPunit is generating output
+        $m = new Project();
+        
+        $response = "";//$m->returnZipFile();
+        $expected = "";
+        $this->assertEquals($expected, $response);
+    }
+
+
+}