+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All Notable changes to `leech:codemonkey` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## [0.0.5] - 2017-07-11
+## [0.1.0] - 2017-07-16
 
 ### Added
-- Made Project::clearTempFolder Public
+- Working Unit Tests
+- Examples for use in Unit Tests
+- Started Documentation
 
 ### Deprecated
 - Nothing
@@ -20,6 +22,13 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Security
 - Nothing
+
+
+## [0.0.5] - 2017-07-11
+
+### Added
+- Made Project::clearTempFolder Public
+
 
 ## [0.0.1] - 2017-07-09